backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3
openai==1.12.0
python-dotenv==1.0.0 